+++ sklearn/svm/__init__.py
@@ -12,7 +12,7 @@ The :mod:`sklearn.svm` module includes Support Vector Machine algorithms.
 
 from .classes import SVC, NuSVC, SVR, NuSVR, OneClassSVM, LinearSVC
 from .bounds import l1_min_c
-from . import sparse, libsvm, liblinear, libsvm_sparse
+from . import libsvm, liblinear, libsvm_sparse
 
 __all__ = ['LinearSVC',
            'NuSVC',
@@ -23,5 +23,4 @@ __all__ = ['LinearSVC',
            'l1_min_c',
            'liblinear',
            'libsvm',
-           'libsvm_sparse',
-           'sparse']
+           'libsvm_sparse']