+++ sklearn/svm/tests/test_sparse.py
@@ -203,28 +203,6 @@ def test_sparse_realdata():
     assert_array_equal(clf.dual_coef_, sp_clf.dual_coef_.todense())
 
 
-def test_sparse_svc_clone_with_callable_kernel():
-    # first, test that we raise a value error for "sparse kernels"
-    # this test is only relevant for the deprecated sparse.SVC class.
-    sp = svm.sparse.SVC(C=1, kernel=lambda x, y: x * y.T, probability=True)
-    assert_raises(ValueError, sp.fit, X_sp, Y)
-
-    # Test that the "dense_fit" is called even though we use sparse input
-    # meaning that everything works fine.
-    a = svm.SVC(C=1, kernel=lambda x, y: x * y.T, probability=True)
-    b = base.clone(a)
-
-    b.fit(X_sp, Y)
-    pred = b.predict(X_sp)
-    b.predict_proba(X_sp)
-
-    dense_svm = svm.SVC(C=1, kernel=lambda x, y: np.dot(x, y.T),
-            probability=True)
-    pred_dense = dense_svm.fit(X, Y).predict(X)
-    assert_array_equal(pred_dense, pred)
-    # b.decision_function(X_sp)  # XXX : should be supported
-
-
 if __name__ == '__main__':
     import nose
     nose.runmodule()